backend/requirements.txt
# backend/requirements.txt
# Python dependencies for vn-ai-trader backend.
# Install with: pip install -r requirements.txt

# ─── Web Framework ─────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0

# ─── HTTP / Async ──────────────────────────────────────────────────────────────
aiohttp>=3.9.0
httpx>=0.27.0

# ─── Cache ─────────────────────────────────────────────────────────────────────
redis[hiredis]>=5.0.0

# ─── Data / Analytics ──────────────────────────────────────────────────────────
pandas>=2.2.0
# pandas-ta
numpy>=1.26.0
vnstock3>=0.1.0

# ─── Settings / Validation ─────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# ─── Logging ───────────────────────────────────────────────────────────────────
loguru>=0.7.2

# ─── Notifications ─────────────────────────────────────────────────────────────
python-telegram-bot>=21.0

# ─── Authentication ────────────────────────────────────────────────────────────
pyotp>=2.9.0

# ─── Database ──────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# ─── AI SDKs ───────────────────────────────────────────────────────────────────
anthropic>=0.25.0
google-generativeai>=0.5.0

# ─── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
